# server/swagger.yaml
openapi: 3.0.3
info:
  title: Nau mai, welcome to Feed the Need
  description: The API is for use with the website [NauMai](http://nau-mai.co.nz) or internal staging at [localhost:3333](http://localhost:3333).
  version: 0.0.1

servers:
  - url: http://localhost:5123/api
    description: Internal staging API server for testing
  - url: http://feedtheneed.click/
    description: Main API server

paths:
  /listings:
    post:
      summary: Creates a new listing.
      consumes:
        - application/json
      produces:
        - application/json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schema/NewListing"
              type: object
              properties:
                donor_id: integer
                title: string
                description: string
      responses:
        "201":
          description: OK - Listing has been listed
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
        "404":
          $ref: "#/components/responses/NotFound"
      tag:
        - listings

  /listings/{listingsId}:
    get:
      summary: Gets a listing by ID.
      parameters:
        - in: path
          name: listingsId
          schema:
            type: integer
          required: true
          description: Numeric ID of the listing to get
      responses:
        "200":
          description: your selected listing
          content:
            application/json:
              schema:
                $ref: "#/components/schema/Listings"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  responses:
    BadRequest:
      description: Bad Request - missing parameter/s
    Unauthorized:
      description: Unauthorized
    ServerError:
      description: Server Errors
    NotFound:
      description: Not Found
  schema:
    Listings:
      type: object
      required:
        - donor_id
        - title
        - description
      properties:
        id:
          type: integer
          primary: true
        donor_id:
          type: integer
        title:
          type: string
        description:
          type: string
          example: I have a bountiful amount of cabbages, that I am willing to donate.
        date_created:
          type: date-time
    NewListing:
      type: object
      required:
        $ref: "#/components/schema/Listings/required"
      properties:
        donor_id:
          $ref: "#/components/schema/Listings/properties/donor_id"
        title:
          $ref: "#/components/schema/Listings/properties/title"
        description:
          $ref: "#/components/schema/Listings/properties/description"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      in: header
      bearerFormat: JWT

security:
  - bearerAuth: []